+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/log/Logger.kt
@@ -15,14 +15,4 @@ package slatekit.common.log
 
 object Logger {
 
-    fun parseLevel(level: String): LogLevel {
-        return when (level.trim().toLowerCase()) {
-            "debug" -> Debug
-            "info"  -> Info
-            "warn"  -> Warn
-            "error" -> Error
-            "fatal" -> Fatal
-            else    -> Debug
-        }
-    }
 }
